Better specify the conditions of the challenge
Applicants usually think that the answer is just a matter of manipulating
firewall rules on Server B, allowing the client to connect to the web port.
The premisses were not making clear enough that Server B cannot be reached
from the client at all.

Also, adding some comments to clarify that a temporary solution is
the answer, and not some big redesign of the architecture.